dieroll working as expected still need to write more tests
